CONN-681: Changed tinyint(1) to number(1,0) for use with Oracle DB

+++ Product/DBScripts/nhincdb/nhincdb_oracle.sql
@@ -78,7 +78,7 @@ CREATE TABLE nhincuser.document (
   DocumentUri varchar2(128) default NULL,
   RawData BLOB,
   Persistent number(11) NOT NULL,
-  OnDemand tinyint(1) NOT NULL default 0,
+  OnDemand number(1,0) NOT NULL default 0,
   NewDocumentUniqueId varchar(128) default NULL,
   NewRepositoryUniqueId varchar(128) default NULL,
   PRIMARY KEY  (documentid)
